Wrote my own MIF converter. Matlab was not functioning correctly.

+++ utils/miffilegen.m
@@ -15,11 +15,11 @@ fprintf(fid,'CONTENT BEGIN\n');
 count = 0;
 for r = 1:rows
     for c = 1:cols
-         red = uint16(img(r,c,1));
-         green = uint16(img(r,c,2));
-         blue = uint16(img(r,c,3));
+         red = img(r,c,1);
+         green = img(r,c,2);
+         blue = img(r,c,3);
          color = red*65536 + green*256 + blue;
-        fprintf(fid,'%4u : %24u;\n',count, color);
+        fprintf(fid,'%4u : %u;\n',count, color);
         count = count + 1;
     end
 end
